DAT-972-importer-errors
