+++ cmdb/database/utils.py
@@ -17,6 +17,7 @@
 List of useful functions for the database
 """
 import re
+import uuid
 import logging
 import calendar
 import datetime
@@ -28,20 +29,14 @@ from bson.objectid import ObjectId
 from bson.timestamp import Timestamp
 from bson.tz_util import utc
 
-try:
-    import uuid
-
-    USE_UUID = True
-except ImportError:
-    USE_UUID = False
-
 from cmdb.framework.rendering.render_result import RenderResult
 from cmdb.cmdb_objects.cmdb_dao import CmdbDAO
+from cmdb.search.search_result import SearchResult, SearchResultMap
 # -------------------------------------------------------------------------------------------------------------------- #
 
 LOGGER = logging.getLogger(__name__)
 
-_RE_TYPE = type(re.compile("foo"))
+RE_TYPE = type(re.compile("re.Pattern"))
 
 ASCENDING = 1
 DESCENDING = -1
@@ -83,8 +78,9 @@ def object_hook(dct: dict):
     if "$maxKey" in dct:
         return MaxKey()
 
-    if USE_UUID and "$uuid" in dct:
+    if "$uuid" in dct:
         return uuid.UUID(dct["$uuid"])
+
     return dct
 
 
@@ -99,6 +95,12 @@ def default(obj):
     if isinstance(obj, (CmdbDAO, RenderResult)):
         return obj.__dict__
 
+    if isinstance(obj, (SearchResult,SearchResultMap)):
+        return obj.to_json()
+
+    if isinstance(obj, bytes):
+        return obj.decode("utf-8")
+
     if isinstance(obj, ObjectId):
         return {"$oid": str(obj)}
 
@@ -108,19 +110,17 @@ def default(obj):
     if isinstance(obj, datetime):
         if obj.utcoffset() is not None:
             obj = obj - obj.utcoffset()
+
         millis = int(calendar.timegm(obj.timetuple()) * 1000 + obj.microsecond / 1000)
         return {"$date": millis}
 
-    if isinstance(obj, _RE_TYPE):
+    if isinstance(obj, RE_TYPE):
         flags = ""
         if obj.flags & re.IGNORECASE:
             flags += "i"
         if obj.flags & re.MULTILINE:
             flags += "m"
-        return {
-            "$regex": obj.pattern,
-            "$options": flags
-        }
+        return {"$regex": obj.pattern, "$options": flags}
 
     if isinstance(obj, MinKey):
         return {"$minKey": 1}
@@ -134,7 +134,7 @@ def default(obj):
     if isinstance(obj, Timestamp):
         return {"t": obj.time, "i": obj.inc}
 
-    if USE_UUID and isinstance(obj, uuid.UUID):
+    if isinstance(obj, uuid.UUID):
         return {"$uuid": obj.hex}
 
-    raise TypeError(f"{obj} is not JSON serializable")
+    raise TypeError(f"{obj} is not JSON serializable - type: {type(obj)}")
