+++ cmdb/framework/models/category_model/__init__.py
