+++ cmdb/interface/gunicorn.py
@@ -23,18 +23,20 @@ from gunicorn.app.base import BaseApplication
 from cmdb.database.mongo_database_manager import MongoDatabaseManager
 
 from cmdb import __MODE__, __CLOUD_MODE__
-import cmdb.process_management.service
+from cmdb.process_management.service import AbstractCmdbService
 from cmdb.interface.net_app import create_app
 from cmdb.interface.docs import create_docs_server
-from cmdb.interface.rest_api import create_rest_api
+from cmdb.interface.rest_api.init_rest_api import create_rest_api
 from cmdb.utils.system_config import SystemConfigReader
 from cmdb.utils.logger import get_logging_conf
 # -------------------------------------------------------------------------------------------------------------------- #
 
 LOGGER = logging.getLogger(__name__)
 
-
-class WebCmdbService(cmdb.process_management.service.AbstractCmdbService):
+# -------------------------------------------------------------------------------------------------------------------- #
+#                                                WebCmdbService - CLASS                                                #
+# -------------------------------------------------------------------------------------------------------------------- #
+class WebCmdbService(AbstractCmdbService):
     """CmdbService: Webapp"""
 
     def __init__(self):
@@ -86,32 +88,38 @@ class WebCmdbService(cmdb.process_management.service.AbstractCmdbService):
         """ignore incomming events"""
 
 
+#CLASS-FIX
 class HTTPServer(BaseApplication):
     """Basic server main_application"""
 
     def __init__(self, app, options=None):
         self.options = options or {}
+
         if 'host' in self.options and 'port' in self.options:
-            self.options['bind'] = '%s:%s' % (self.options['host'], self.options['port'])
+            self.options['bind'] = f"{self.options['host']}:{self.options['port']}"
+
         if 'workers' not in self.options:
             self.options['workers'] = HTTPServer.number_of_workers()
+
         self.options['worker_class'] = 'sync'
         self.options['disable_existing_loggers'] = False
         self.options['logconfig_dict'] = get_logging_conf()
         self.options['timeout'] = 120
         self.options['daemon'] = True
+
         if __MODE__ in ('DEBUG','TESTING'):
             self.options['reload'] = True
             self.options['check_config'] = True
             LOGGER.debug("Gunicorn starting with auto reload option")
+
         LOGGER.info("Interfaces started @ http://%s:%s",self.options['host'], self.options['port'])
         self.application = app
         super().__init__()
 
 
     def load_config(self):
-        config = dict([(key, value) for key, value in self.options.items()
-                       if key in self.cfg.settings and value is not None])
+        config = {key: value for key, value in self.options.items() if key in self.cfg.settings and value is not None}
+
         for key, value in config.items():
             self.cfg.set(key.lower(), value)
 
@@ -120,12 +128,16 @@ class HTTPServer(BaseApplication):
         return self.application
 
 
+    def init(self, parser, opts, args):
+        raise NotImplementedError()
+
     @staticmethod
     def number_of_workers() -> int:
         """TODO: document"""
         return (multiprocessing.cpu_count() * 2) + 1
 
 
+#CLASS-FIX
 class DispatcherMiddleware:
     """TODO: document"""
