+++ cmdb/utils/helpers.py
@@ -45,7 +45,7 @@ def load_class(classname):
     #TODO: check if this regex is correct
     pattern = re.compile("(.*)\.(.*)")
     match = pattern.fullmatch(classname)
-    LOGGER.debug(f"[load_class] pattern match: {match}")
+    LOGGER.debug("[load_class] pattern match: %s", str(match))
 
     if match is None:
         raise Exception(f"Could not load class {classname}")
