+++ tests/fixtures/fixture_rest_api.py
@@ -17,7 +17,7 @@
 import pytest
 from cmdb.interface.cmdb_app import BaseCmdbApp
 
-from cmdb.interface.rest_api import create_rest_api
+from cmdb.interface.rest_api.init_rest_api import create_rest_api
 from tests.utils.flask_test_client import RestAPITestClient
 # -------------------------------------------------------------------------------------------------------------------- #
